#507: Add display for preview view mode for every file type noting gets rendered using file_link.
When a file is rendered using file_link, it's unselectable in the media browser. File Entity uses file_link as a fallback in case a file type has no displays that render anything for a given view mode. By providing some display, we prevent the fallback from occurring and the Media Browser is usable again.