Implemented OSMigrateImageFIeldHandler #523
Gallery migrate works again!

We added a custom field handler for files so we could add og_group_ref.  To activate it we disabled the default one.  Even though the default image field handler was still enabled, it wasn't being used becuase its parent handler was disabled.  To get around this I rewrote it as OSMigrateImageFIeldHandler and stuck og_group_ref on there.