Migrate shield, oembed fix.  #351 #424
Earlier I assumed I knew the path to vsite_shields.  That was a bad assumption.  Had to write a recursive scandir to find it for me.

Had a fubar query for migrated oembed files.  Unsure how I wrote it in the first place, let alone assumed it would work enough to commit it.  It's not even the right table.  But now it's fixed.