imagecrop migration recycles d6's thumbnail when available.  #535
Wish I did this in the first place.  When both the original file and thumbnail are available, the imagecrop migration field handler adds both files to the database, calls file usage_add to say that one is the crop of the other, and returns the fid.  The image is only cropped during migration if a thumbnail cannot be found.