Fixed crop for person migration.  #535
I can see now why zendstudio tells me not to set variables in conditionals.  Setting the variable worked initially, but a couple commits ago I added an "&& $gid" to make sure we also had a group to set.  Well, now the variable was being set to the boolean result, not to the function it was calling at first.  Parentheses to the rescue.