Fixed media_gallery row/paging.  #523
Not sure when the change happened, but the render array for the gallery node was built before os_image_gallery saw it and passed it on to media_gallery (which was where it was supposed to limit the files present based on rows/columns).  I kept the rendered data and stripped out the files media_gallery didn't want.  Rows/paging works again.