Announcement -> news migration.  #563
Changed the class to create news nodes instead of announcement nodes.
Updated the menu links in spaces overrides.

After doing those, I found the news feature was disabled in my vsites.  Updated spaces_overrides_variables, so that migrating the spaces_features setting will use os_news instead of scholar_announcements.  That change applied to toher places that refer to the module by name, but I don't expect that to be a problem.

So far so good.