Fixed inline embed migration.  #534
Media embeds defined by embed tags weren't migrating correctly.  Added function to rebuild those tags.  The new version of media required me to change what I was passing to the file entity to get it to work as an oembed.  And we discovered that the embedly api key variable's name changed, so I spent a good portion of the day wondering why I was blocked from embedly (that change has been accounted for in migrate).

There are some files in the DB that won't work for embedly, but in all the cases I saw they were in older revisions of their nodes.  Looks like someone tried to make an embed and when it failed they went at it another way instead.  I couldn't find any instances of embeds that worked in 6 but not in 7.