Fixed missing gallery node.  #598
Slideshow box was migrating correctly, but was pointing to a gallery node that did not migrate.

The node wasn't migrating because its filter format was incorrect.  When it joined the filter format table to get the name of the filter, there was no filter of that id so the node was ignored.  I changed it to a left join and provided a default filter format.  Now the gallery node migrates.