#646: Wrap wysiwyg files in span instead of div, and apply attributes to that span.
THe problem wasn't that the file was being wrapped in something. The problem is that divs can't be in p's. It's invalid markup. The browser will pop them out to make it valid, then render that. As a result, this would break float formatting that is common with images. To fix this, I changed the div to span, and applied style attributes to the span. They float correctly now.