Fixed os_boxes_node -> os_boxes_html inline media.  #643 #623
The os_boxes_node to os_boxes_html migration combined with a known bug in drush was the source of this error.  Basically drush didn't have the base_domain.  When inline images and files were rendered to be saved as html, they were given an absolute url using http://default.  The drush bug is what made this obvious, but even if that were fixed or if we were migrating through the web UI, sites with custom domains would have the wrong url.

The solution was to use media tags instead of the links provided by media.  When the box is rendered, the media tags will be swapped out for html.  As long as they're stored in the box, there's no reason to render them now.

Instead of tokenizing, I let them render.  Then I got a list of the media tags in the body and did check_markup (with the filtered html filter) on each of those to get its HTML.  I swapped out the HTML for the original tag and saved teh resulting markup.