Forgot to update sv list migration for content type changes.  #663
Latest announcement was being migrated, but the content type never changed from announcement to news.  The old announcmeents block wasn't being displayed and we got the default one instead.

Context should need an update too, but for now I'm happy to delete the extra default box if this one fails to replace it.