os_migrate has admin setting to restrict vsite.  fixes #685
Added a setting to migrate's admin page that takes a comma separated list of vsite ids.  Updated affected files to use those ids isntead of all content.

Only did this for queries that already have site id.  The url_alias table doesn't, so it couldn't be restricted without adding joins.  Not worth it.