#133: Replace Individual region permissions with a single permission and a variable.
Replaced the 'Edit Region region' permissions with two permissions: Edit locked regions and Managed locked regions. Edit lets you do things to the locked regions, and Manage lets you set which regions are locked. Which regions are locked is stored in a variable, although there's no UI to set this yet. The permission is only checked when a region is locked, otherwise it falls back to spaces_access_admin.