#719: Change unset() to #access = false
I glanced through comment sources quickly, and it seems that it expects the comment settings for every node type in the same form. If a node type isn't in a given form, then its settings are bulldozed. Changing the unset() call to set access = false instead seems to fix the issue.