Fixed presentation files.  #722
File paths are fixed.  Files use old description as titles.  This means that Gary's presentations will display as "slides" instead of "talk-$date-_1.ppt".

Had some trouble updating old code to work as desired, so I gutted it and used the code from node abstract.  It wasn't quite generic enough to work as it was written, so I copied and fudged it instead of risking breaking the generic funciton.

Note: while most file subfields can take arrays, the one for filename apparently doesn't.  When it was present I kept running into trouble with migrate expecting arrays of files to be strings of file.  When that subfield was omitted, the arrays worked fine.  Couldn't figure out why, so I renamed the files after creating presentation nodes.