#727: Push content region check later in layout form.
If a context has a widget in a content region, we assume the content regions should be active. Any context that uses the content regions should have the main content widget in a content region, to simulate page appearance with no context. A special exception is made for the front page.