User migration suppresses email and mailchimp.  #827
User migration sets a flag before starting, indicating that user migration is in progress.  Flag is deleted upon completion.

When this flag is active and the suppress email setting is enabled (which is the default), hook_mail_alter blocks the welcome message and hook_module_implements_alter removes os_mailchimp's implementation of hook_user_insert.