Fixed SV list pager.  #836
This wasn't a migrate issue but a pager one.  The pager wasn't working correctly on most content types and wasn't even trying to work for biblio.

For it to work with biblio, I needed to initialize the pager in the ajax callback function.  Doing so involved knowing the total number of nodes in advance.  To do this I separated out the portion of render that fetches nids into a separate get_nids() function.  That function handles the EFQ (or whatever pseical case is used for biblio) and just returns nids.  Render() calls get_nids() and sets up the display of those.  Not much of the code actually changed, but it moved around lot.

I tested the following displays of the pager.  Announcement initial page load, announcement after paging, biblio initial page load, biblio after paging.  Those all work.