#727 #826 #825: If the custom context is found, but has no blocks, return an empty array from os_layout_get().
It was returning false before, but we weren't checking that value against anything. We just assumed it was an array and moved on.When the custom context was created, most of them don't have a block reaction yet. If they didn't have a block reaction, os_layout_get would return false. Now we only return false if the context doesn't exist and no module claims it with hook_os_custom_context.