SV list migraiton.  Use presetnation sorting. #838
Wrote a quick fix.  During recent presentations migration, get the vsite's presentation dates.  If all presentations have a date, switch to presentation date sorting.

This almost fixes the original box's problem.  That site has 92 presentations, 1 of which is missing a date.  Not sure if I should call that close enough, or if we should just make an exception for htat one vsite.