#727 #825 #826: Pass more than one context to layout form.
Previously, the context form only accepted one context, and os_public was hard-coded in. It's now possible for some pages to have 3 contexts active, so the layout form needed to be updated to handle this case.

Also, with the change to context that allows modules a last chance effort to define a context, we no longer need to preprocess the page to add a new context to the Layout link. It's just there.