Recent image widget filters vsite.  #868
Had an entity condition that should have been a field condition.  Now images are filtering.  Not all migrated images are making it in though.  Since gallery images are attached to a node, I'm not sure they have an og_group_ref of their own, since that's a field on the node.  This might require a full query intead of an EFQ.