Migrate updates HTML of node boxes to mimic jcarousel  #867.
Problem with this approach: the divs don't work under filtered html.  Setting the widget to full html fixes that, but I'm not sure that's the best answer here.

Todo: Post in a widget boxes weren't the only source for this problem.  Bio boxes worked the same way.  Make sure they work as well.