#877: Pages as book type. 'Add Subpage' link on every page. No book is made until child page is created. After that, core takes over completely.
Most of the work is done in hook_node_submit, which only does things the first time a subpage is added. It creates the book out of our parent page, which also creates a menu, and a menu link. Then it passes all that into the subpage node. From there, book handles the rest.