#878: Add contextual link to section nav widget.
This contextual link goes to a drag and drop form to manage the outline of the section. It's a form implemented in core, but is in the admin area. I implemented a simple menu entry to call this form directly, which also let me do whatever I want with the menu router, such as change the path and make it usable in a contextual link.