Jcarousel html fix.  #867 [status: needs qa]
Not sure why, but when I was migrating under drush the order of html attributes kept changing order.  This made the regex approach tedious, so I opted to DOM manipulation isntead.  No new logic here, but I'm unfortunately not as familiar with php dom manipulation as I'd like to be so this was slwoer than expected.