Fixed post in a box migration for biblio nodes.  #939
Started out changing structure just like I had for jcarousel.  Instead went for full_html setting.  That fixed all the js and display issues, so now the biblio citation is identical to the one you'll see in publications.  It is changed from what was on the original site, but is consistent with d7 biblio.  If this is incorrect, we should make the change in biblio.