Software release migration.  File, field fixes.  #1003
There was a typo in a field mapping.  An extra : appeared in place of an _.  This gave it two subfields.  Hence the missing $field error.

File source path was messed up.  Thought this was working after the latest change.  It's possible it was still problematic but I was losing the error in the stream of notices.

Still getting the image_dimensions notice in both release and projects.