Fixed galleries not listed in slideshow.  #1066 [status:needs qa]
This wasn't a migrate problem after all but a slideshow box one.  The slideshow box was populating its list of slideshows with a view.  Because I was being cheap when I wrote that section, it was recycling the view used later on to display images.  It got a list of all the images in slideshows and put their nodes into a list.

Problem was that there was a pager so it was only getting the first ten items.  That's first ten files.  Which equated to two slideshows.  Replaced the view with an EFQ.  Now it gets all slideshows in a vsite.