Fixed large inline images being shrunk.  #1089
Images were being shrunk because the biggest they were using too small of a media filter.  #1095 provides a larger filter for these images.  This commit lets them use it.

Could not fix the class issue.  Adding the class to the media tag did not get the class onto the page.  This is a media problem, not a migrate one.