Changed file path matching order.
Unmatched paths matched this time, but were not replaced correctly.

Specifically "../file.ext" ended up "node/files/file.ext".  I think an earlier path was matching the file.ext and stripping it before we got to the ../ path.  Moved those to the top because this won't work if a ../ is left in the url.