Unlisting a file from a node shouldn't delete the file.  #1177
Original version did file_usage_add, before unsetting the file from $node->field_upload.  The logic was to increase the usage counter, then it would decrement automatically when the node was saved.  I think what was happening was that all the usages were on that one node, so it removed the file.

New version counterintuitively does the opposite.  file_usage_delete removes the link from the node to the file, but the count parameter is set to 0 so that none of the counts are updated.  Now the file is no longer linked to the node when the node file is stripped out and the node is saved so the count remains intact and the file isn't deleted.

Tested locally.  Listed files show up on node and cp/content/files.  Unlisted files only show up in cp/content/files.  Need to test thoroughly on staging.  I do NOT want to fix this in an update hook again.