Fixes publication sorting.  #1218 [status: needs qa]
Publications were having their sort order completely reset so we could use year>date>title instead of the default year>title.  This was happening even when a sort order was specified.  I checked if the query had a sort order and only reset the order when there wasn't a deliberate sort query.