Migrate filename fix.  #1177
More updates to the code that updates body links with the new urls of attached files.  Not positive it's ready to go, but it is ready to run outside of my local environment.

Replaced paths use the base domain and a custom domain.  Probably not necessary, but it's possible a user manually saved a link in a node body and then added a custom domain.

Also checked the legacy DB for filefield paths to replace.  That's kind of important.

Saved a success message.  This should help us QA afterwards.