Fixed vsite imagecrop file_copy.  #1177
Was checking for a path and defaulting to "public://" when no path was provided.  Used isset for the check, so when the path was '', that got used.  Whoops.

Also found a potential problem in how the paths are generated.  public://$purl/files can give you public:///files if no $purl is available.  I don't think that will be the case, but fixed that one anyway.