Files link updater was missing some files.  #1177
@cfirth1005 found some files that didn't have their paths updated.  This fixes those files.

The function that updates paths checks for attached files and loops over those to get their new paths.  Another function was removing unlisted files so they wouldn't be displayed in the node.  When the files were removed first, the link updater didn't have the files anymore so it didn't work.  Now they're in the correct order.