Fixed sv list's biblio type filtering.  #1348
The sv list used to have a checkbox for selecting if all publication types should be shown or just a subset of them.  When clicked the checkbox was supposed to show/hide the list of publication types, but this didn't work so well with Drupal form states since the checkbox itself was only visible when biblio content was selected.  The checkbox was removed, but its setting was still being checked before biblio types were filtered.

The fix was to stop checking for that setting.  New sv list boxes get the full list of biblio types all enabled.

The only problem with this is that the existing sv lists won't have any biblio types checked if they were kept default.  Currently working around that by only filtering types when there's at least one type selected (the assumption being that an sv list with 0 biblio types default, and not deliberate).  An update hook would be a more correct fix though.