Began OS Migrate docs.  #767
known_issues.md is mostly complete.  It lists the things that didn't migrate so well that needed manual fixing later.  I went through all the old tickets to find issues for this.

readme.md is about half done.  It details configuring and running migrate, but glosses over the rest of the setup.  Not sure how much detail is needed here.  This document will benefit from review by someone who hasn't already been through the migration process a couple hundred times.

development.md needs the most work.  I provided links to some of the posts I used for reference because I'd like to keep the scope of this to extending OS's implementation of migrate.  No need to reinvent the wheel with intro to migrate content.