#1261: Redo how we iterate over files.
Previously, the update was missing over 1000 files, which is probably a result of the moving target we were trying to hit. To handle it, I got all the fids we would need to update at the start and stored them in the sandbox. This gave us a non-moving target, so each pass could carve out a chunk of this dataset and process it.