Media gallery takes embedded media correctly.  #741
Media gallery wasn't listening to file settings because it was hard coded to use images.  Fixed it to use whatever types the file field is configured to use instead.  Couldn't do this outside of the contrib module unfortunately.

Started writing an update hook to update the file field settings.  Need to double check that it's actually necessary.