Media gallery shows preview image of movies.  #741
Instead of having a gallery full of iframes to flash videos, gallery now populates itself with thumbnails of those videos.  Clicking on a video to view in the preview lightbox or the video's detail view shows the full flash embed.

Overriding media_gallery's theming was a little more circuitous than intended.  At the bottom I have a theme function and an implementation of hook_theme to register it.  I couldn't just use that theme function though, because oembed provides #pre_render functions that override whatever theme I select.  So I added my own pre_render function to change it back.  The file_field_inline theme function can't tell if it's displaying in a gallery or a single item view, so I added a theme preprocess function that sets a flag when we're displaying a thumbnail.  Much more convoluted than intended.

Also had to change the file's #cache keys.  If thumbnails are rendered differently than full or lightbox view, they shouldn't be stored in teh same cache.