Sortable slideshow widget conifg.  #1331
Widget has sortable table.  Need to update display of it to show image instead of nid.

Problem.  How to tell the view to use a custom order?  Can't just sort by nid since that's not correct.  The weight is stored in the box options, not in the node.  Working on this next.