Slide show displays in order.  #1331
Not totally please with how I did this.  Couldn't find a way to specify a custom sort through views options, so I altered results during views_pre_render.  I'd like to find a cleaner way to do this, but this one is working for now with only one hook.  Performance-wise, this isn't a big list of images.  It's a slideshow.  It should have fewer than a dozen images.