#1394: Get real original context.
Spaces was using context_load to get the 'original' value of the context. The same context_load that spaces alters to override the context. It was getting overrides and thinking they were original. Because of this, if a context override was to be saved with no changes made, it would be deleted entirely instead.