Slideshow -> HC slideshow.  Old slideshow re-enabled.  #1331
The old slideshow is no longer out of commission.  Some users may still want it.  To differentiate them, the new one is called HC slideshow.  Anyone working with a HappyCog theme will know what that means.  Left the ctools hook commented out so we could deprecate the old slideshow again.

Fixed a missing isset in the old slideshow while I was in there.