Fixed change note settings in os_publications.  #1443
That was bigger than expected.  In d6 we used to have another menu path for loading app settings in a popup or without js.  That was never implemented.  I added the menu item and set up a callback to return ajax (or route you to the cp).  Updated the submit function that closes the modal.  While I was in there I also added a little ajax to update the note settings after the form saves.  It was that or reload the page.