hwpi_basetheme overrides os_slideshow controls.  #1557
theme_registry_alter adds our slideshow_controls_text template.  Also needed a hook_preprocess to add the rendered pager to the var passed to that template, since views_slideshow kept them separate.  All the template itself does is replace the pause button with the pager, so that part was dead simple.

Not totally sure this is working.  It runs through the debugger correctly, but there's another issue going on.  My controls don't show up.  They fade in and then fade out.  Dan suggested pulling in the 3.2 version of basetheme css, but that made no difference.  (Those updates are included here as well).