Slideshow width is automatic.  #1557
Removed slideshow size control.  Slideshows now assume it was 400px for sidebars or 1000px anywhere else.  I started using the currnet theme's info file to determine the width, but I couldn't tell if that was reliable.

Sizes are still being removed by a stupid old str_replace.  This is how we changed the sizes originally.  Now we're just changing them to 100%.  I'd like to update the view instead, but for now this was trivial.

@dancabral, I'm sorry for what this means for css.  A slideshow whose region has been shrunk with sidebars is going to be smaller than expected and its description box will suffer.  ![screenshot](http://i.imgur.com/0yildDq.png)