Multiple slideshows work on same page  #1557
(Not sure this use case is valid, but it's the principle of the matter.)

I had been using box deltas so we could enumerate over all the slideshows on a page.  However drupal_add_js was munging the deltas because they looked like non-associative array indices.  Changing them to strings fixed that, so I went ahead and used the box selector as an id for each.

The other problem was the events.  All the events were pointing to the last slider, since that was the value that was left after I looped over them all.  Instead of adding more closure shenanigans, I had each of the triggered elements navigate up the hierarchy till they got their parent .flexslider div, which has access to flexslider's api functions.  Next, prev, pause, and pager all work.

The pager could probably use a .active class to indicate which slide is being viewed.