Fixed slideshow fieldset, unpublished behavior.  #1557
The published checkbox was the only thing in the fieldset on slideshow_slides so I removed the fieldset.  Made status=0 work as intended.  Added an indicator to the widget when a slide is unpublished.  I figured someone would click that check box to see what it did and then forget about it, only to complain three months later that an image was in the config but not the slides.