#1716: Render a placeholder image in wysiwyg.
Media used to take care of this before. If a formatter returned nothing for a wysiwyg, it would generate a placeholder image. Now the default behavior is to generate a link to the file instead. This is useless for iframes, so I had to render something. With that fixed, the rendering issue on node view fixed itself.