Hide edit, outline tabs.  #1621
Removed links that weren't view or devel from the tabs.  Adaptive uses a different variable and loads primary local tasks on its own, so I had to change it at the theme level as well.

Edit was already in the contextual so no change there.

Should devel be in the contextual?  It only really affects us.