SV list uses api function to get paths to content type.  #1167
The SV list was using two separate arrays of node type => paths to prepopulate the more link when the user changed content type.  I updated it to use `os_get_app_path()` from the os api includes file.  I also had to update that function to return multiple paths when no content type arg was provided.  (The function was only used once according to grep and this change won't break that usage).  Also fixed a small javascript problem while I was in there.

No idea what the vsite comments were about.  This commit is aimed at the title of this ticket, not the description.