Fixed slideshow imagecrop.  #1557
The hook to change crop size wasn't working in 3.3.  Not sure if it was fragile to begin with or had something to do with media changes.  This fixes @gcaro's problem.

Fixed it.  Re-rolled the imagefield_crop patch and sent it to d.o.

Also found a hardcoded pixel size, so I removed that and used my region size function to get the max size.