Prevent imagecrop hook from breaking non-slideshows.  #1800
My imagefield_crop widget alter hook in os_slideshow wasn't as precise as I thought it would be.  I was expecting boxes_load_box to fail when no widget delta was provided, but instead it gave me ALL the boxes and let the hook try to alter things.  Fixed it so it only does this when there really is a widget delta.