Undid some events changes.  #1637
HWPI themes have a calendar icon that shows the date in teaser mode.  This obviated the needed to include the date by name, so I removed it.  Then I removed it from os_basetheme as well.  But the calendar icon never made it to os, so it didn't make sense to have the date removed there.  On top of that the timezone was buggy.

Updated os_events to use a date format that's short enough to look good in a teaser.  No info is lost on an os theme.  HWPI themes have some redundancy, but it still looks okay.  We'll have to do this in the next release instead.