Fixed slideshow pager events  #1842
Wrapped slideshow init in a .once to keep ajax events from creating more pager buttons.

Started fixing multiple slideshows on same page.  That functionality was working on flexslider but I forgot to switch it when we came back to responsive.  Responsive does offer a namespace argument, but that changes class names, which will require some theming work to fix.  I namespaced slideshows after the first, so it's less borken than it was before, but not quite complete.