Fixed slideshow's slide size.  #1847
I had a miscalculation in how I was figuring out how big a region could be.  I started with the widest possible content area width.  For each sidebar that was active, I subtracted the width of that sidebar.  Problem was, the sidebar widths I was using were the biggest possible because I wanted to know how big to make a slide when it was in a sidebar.  EPS found a theme where the sidebars were considerably thinner than the max allowed.

The fix was to also store the minimum sidebar width and use that when subtracting sidebars from the content area.

This requires a cache clear.