Add menu item takes rel urls, offers new node.  #1958
When add menu item has an internal path that goes nowhere, it now offers to create a new node using that path.  I didn't want to add a new page for dealing with this, so the invalid path error message now presents a select box and an explanation.  Getting that form item to show up was trickier than expected.

While I was doing this I ran into a case where form errors weren't being returned.  Explicitly returned them as part of the ajax command already being sent.

This isn't quite finished yet.  It saves everything that needs to be added to the node form to the wizard cache, but I'm getting og_vocab errors on node/add, so I haven't gotten to deal with the rest of it yet.