media_gallery - refixed hijacked taxonomy page.  #896
media_gallery oversteps its bounds and hijacks a taxonomy term page to display its collections.  This ended up taking over /galleries as well, which is why our view wasn't viewable.  I'd fixed this in a previous version, but it wasn't applied in the recent update.

Re-fixed it.  Rerolled patch and sent that in.

So far this looks good, but I'm still waiting for my prod files to sync so I can test it for sure.