Fixed sv list more link paths.  #1167
Image was a special case.  It's not a node, so there was no os_app_info hook deifning its path.  Pointed it to galleries.
Page was erroneously pointing to pages.  Removed that from the pages app info hook.
All has no single page to point to.  Set it to blank the textfield so users can enter it themselves.