Fixed os_get_app_path.  #1167
There was a logic problem in os_get_app_path.  That function gives you an app's path from hook_os_app_info, or all app paths if no app provided.  But if an app didn't define a path or its path was '', returned the array of all apps.  I fixed it to return false instead.

Cleaned up sv list javascript.  It swaps things correctly now.  I ended up removing the functionality to preserve the original more link.  If someone has an sv list with a custom path, they're probably not changing the content type and changing it back.  That part of the script wasn't necessary and was error prone when ajax reloaded the event.

Also made the swapping a little more error proof.  Content types that aren't given a path default to ''.