Slight fix for media_gallery update hook.  #896
Told the update hook not to run if the new field already has a table.  This isn't the case in prod, but if someone installed between now and when the field changed, their DB would have the new field and this update hook should not be run.

Also updated a comment.