Hide slideshow when media_gallery is disabled #2002
Slideshow widget behavior now matches the description.  When media gallery is disabled, the slideshow turns off.

I would love to have been able to disable the save button on the widget creation form, but that's part of os_boxes and it would be hacky to remove.  Also it wouldn't help if someone enabled media gallery, saved a slideshow, and disabled media gallery.