Cleaned up cp changes, fixed publications note link.  #1443
Cleaned up my changes from a previous commit.  Left the change that let the settings page work even if it wasn't loaded in an overlay.

CHanged the publications note settings link to point to cp/build.  I realize the ticket said to tell people to go to the publications customization section of cp/build, but as long as that text was there I didn't see a reason not to link it.  No special menu entries this time though.