Fixed media_gallery descriptions.  #2192
Bug in os_files was breaking the form element that took media_description.  I removed it.

While the bug was active, several descriptions were given an incorrect format.  Both the description and format were set to the first character of the user's description text.  Wrote an update hook to correct this.