Copied un-makeable files from 3.x.  #2100
Made a contrib/ outside of openscholar/modules.  It's full of the modules that didn't build right, as we have them in 3.x.  After building, they get copied into contrib.

The idea is to remove these and add them to the make file one at a time as we figure out their versions and get their patches into d.o.