Fixed output bug, wrote help docs.  #1550
This is nearly complete.  ac-domain-* is missing.  I haven't been able to get the cloud commands to auth from drush_invoke_process.  I'd like to move this in its current state to dev, so we can test it there.
Some path overriding may be necessary.  Not sure until we run on dev