Fixed jquery.cycle library.  #2100
Our build was fetching jquery.cycle.all.2.88.js.  The path to the library was hardcoded to jquery.cycle.all.js.  Drush make lacked options for renaming downloaded library files.  I tried playing with hook library to see if its version support would help, but it didn't.

Gave up on that.  Instead linked to raw download of older revision of the file, which still had right file name.  I think that's the right way to do this in the make file for library downloads.