Renamed mailchimp var.  #2223
Mailchimp serves two purposes.
1. Provides a box so vsite owners can have a sign up for my list link.
2. Automatically subscribes new user accounts to our mailing list.

For some odd reason, when I set this up both of those used the same os_mailchimp_api variable name to store the key.  Spaces didn't like that.  It thought the two were the same variable.  This meant that if you set the global api key and then set a vsite's api key to the same thing, the vsite wouldn't store an api key.  Spaces thought it was using the default.

I think this is an edge case.  It mostly applies to testing.  But if we ever wanted openscholar.harvard.edu to have a link to let people subscribe to our list and we used the same list as the auto-subscribe, then we'd run into this.  So I renamed the var and wrote a quick update hook.