Presentation body will show in sv list.  #2489
This is half done.  Body text does not show in teaser mode by design.  It should show in full display mode though.  The presentations template was omitting body text because it was printing it depending on the $page variable rather than $teaser or $display.  That was a relatively easy fix.

Except for one edge case.  When viewing a presentation node, the title is printed outside of the template.  Since the template isn't printing that node's title, it will be missing in an sv list.  I haven't found a way for the template to differentiate between a node on the page and a node in the widget when viewing the node's actual page.

See screenshot.  (Red arrow points to the one missing title) http://i.imgur.com/JGSG2Jb.png