Loosened requirements on drush os-domain-check  #1550
To check if sites.php was configured correctly I was looking for a line that had require_once pointing to the filename (not full path) but the sites.php.inc file.  This was a little too strict, as the actual sites.php sets that file name in a variable and includes the var.  Now I'm just checking if require_once and the file name both exist in sites.php, assuming that if they're there it's deliberate and functional.