Drush domains fixed domains.list #1550
The location of domains.list can be overridden, but it still must be domains.list.  Updated docs to be aware of this.  Updated check function to assert this.

Removed call to drush_file_append_data.  It only appeared once so I replaced it with its source.  It only existed to save 5 lines of typing.