Flattened list of domains  #1550
Our hosting does not differentiate which site is where.  It just needs to know that a list of domains should go to our webroot.

I had been using a single installation's domains list to update acquia's settings.  That would be wrong since each time we added a domain to an installation it would strip out all the installation's multisite brethren's domains.