Used newer version of og patch.  #2371
Cleaner way to fix the queue.  My patch fixed two problems, but Amitai didn't think one of them was a problem the worker should be addressing.  If an item isn't being deleted from the queue, we should fix it so it can delete even if that means the queue doesn't work in the meantime.