Moved inf scroll function.  Applied to taxonomy.  #2680
Publications aren't the only non-view infinite scroll is being added to.  Since os_taxonomy is getting them too, it doesn't make sense to have it be a function of os_publications.  The function to add infinite scroll to a page has been moved to an include in os, accessible with ctools_include.

Added infinite scroll to os taxonomy.

Tweaked how the pager gets shown and hidden.