Prevent domain script from removing original domain #2779
Each of our acquia installations has an acquia-sites domain that cannot be removed.  But that domain shouldn't be part of the domains.list because we don't want it in sites.php or .htaccess.

Since that domain is always the final item on the ac-domain-list, I popped it from the array before making the list of domains to remove from acquia's list.